Refactor components into their own Page directory: Home, About, and Playground
